Minor bugfixes

+++ cleanup.mod.sh
@@ -1,2 +1,2 @@
-rm $OUTPUT
-rm $INPUT
+if [ -e $OUTPUT ]; then rm $OUTPUT; fi
+if [ -e $INPUT ]; then rm $INPUT; fi
